Restructure loops directory
