--- src/anomalib/training/learning_strategies/__init__.py
@@ -1,4 +1,0 @@
-"""Learning strategies for Anomalib models."""
-
-# Copyright (C) 2023 Intel Corporation
-# SPDX-License-Identifier: Apache-2.0
