+++ tools/train.py
@@ -18,7 +18,7 @@ from anomalib.config import get_configurable_parameters
 from anomalib.data import get_datamodule
 from anomalib.data.utils import TestSplitMode
 from anomalib.models import get_model
-from anomalib.training import AnomalibTrainer
+from anomalib.trainer import AnomalibTrainer
 from anomalib.utils.callbacks import LoadModelCallback, get_callbacks
 from anomalib.utils.loggers import configure_logger, get_experiment_logger
